% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_correction.R
\name{get_correction}
\alias{get_correction}
\title{Get correction}
\usage{
get_correction(IVs, lambda, lambda_se, h2_LDSC, h2_LDSC_se, alpha_obs,
  alpha_obs_se, n_exp, n_out, M, MR_threshold, verbose, sn = 100)
}
\arguments{
\item{IVs}{xx}

\item{lambda}{xx}

\item{lambda_se}{xx}

\item{h2_LDSC}{xx}

\item{h2_LDSC_se}{xx}

\item{alpha_obs}{xx}

\item{alpha_obs_se}{xx}

\item{n_exp}{xx}

\item{n_out}{xx}

\item{M}{xx}

\item{MR_threshold}{The threshold used to select strong instruments for MR, should be lower
than 1e-5, \code{default=5e-8} (numeric)}

\item{verbose}{A logical indicating if information on progress should be reported,
\code{default=TRUE}}

\item{sn}{The \code{default=10,000} (numeric)}
}
\description{
Calculate corrected causal effect estimate (SE) and the covariance between
observed and corrected effects, and test for the difference between the two
}
