% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRlap.R
\name{MRlap}
\alias{MRlap}
\title{MRlap - main function}
\usage{
MRlap(exposure, exposure_name = NULL, outcome, outcome_name = NULL, ld,
  hm3, MR_threshold = 5e-08, MR_pruning_dist = 500,
  MR_pruning_LD = 0, MR_reverse = 0.001, s = 10000,
  save_logfiles = FALSE, verbose = TRUE)
}
\arguments{
\item{exposure}{The path to the file containing the GWAS summary statistics for the exposure,
or a \code{data.frame} (character, or \code{data.frame})}

\item{exposure_name}{The name of the exposure trait, \code{default="exposure"} (character)}

\item{outcome}{The path to the file containing the GWAS summary statistics for the exposure,
or a \code{data.frame} (character, or \code{data.frame})}

\item{outcome_name}{The name of the outcome trait, \code{default="outcome"} (character)}

\item{ld}{The path to the folder in which the LD scores used in the analysis are located.
Expects LD scores formated as required by the original LD score regression software.  (character)}

\item{hm3}{The path to a file of SNPs with alt, ref alleles and rsid used to allign alleles across traits
(character)}

\item{MR_threshold}{The threshold used to select strong instruments for MR, should be lower
than 1e-5, \code{default=5e-8} (numeric)}

\item{MR_pruning_dist}{The distance used for pruning MR instruments (in Kb), should be between 10 and 50000,
\code{default=500} (numeric)}

\item{MR_pruning_LD}{The LD threshold (r2) used for pruning MR instruments, should be between 0 and 1
(if 0, distance-based pruning is used), \code{default=0} (numeric)}

\item{MR_reverse}{The p-value used to exclude MR instruments that are more strongly associated with the outcome
than with the exposure,\code{default=1e-3} (numeric)}

\item{s}{The number of simulations used for sampling strategy to estimate the variance of the corrected causal
effect and the covariance between observed and corrected effects \code{default=10,000} (numeric)}

\item{save_logfiles}{A logical indicating if log files from LDSC should be saved,
\code{default=FALSE}}

\item{verbose}{A logical indicating if information on progress should be reported,
\code{default=TRUE}}
}
\description{
Performs cross-trait LD score regression, IVW-MR analysis and provide a correction
that simultaneously accounts for biases due to the overlap between the exposure and
outcome samples, the use of weak instruments and Winner’s curse.
}
\details{
\code{exposure} and \code{outcome} are required arguments.
The input file / data.frame should contain the following
columns (lower or upper case) : \cr
SNPID (rs numbers) should be : \code{rs}, \code{rsid}, \code{snp}, \code{snpid}, \code{rnpid} \cr
CHR (chromosome) should be :  \code{chr} \cr
POS (position) should be :  \code{pos} \cr
ALT (effect allele) should be : \code{a1}, \code{alt}, \code{alts} \cr
REF (reference allele) should be : \code{a2}, \code{a0}, \code{ref} \cr
Z (z-score) should be : \code{Z}, \code{zscore} \cr
N (sample size) should be :  \code{N} \cr
If Z is not present, it can be calculated from BETA and SE. \cr
BETA should be : \code{b}, \code{beta}, \code{beta1} \cr
SE should be : \code{se}, \code{std} \cr

Setting \code{s} to a smaller value is strongly discouraged, it can lead to an innacurate
estimatation of the corrected effect SE, therefore affecting the results of the difference testing
between observed and corrected effects.
}
