% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_LDSC.R
\name{run_LDSC}
\alias{run_LDSC}
\title{Run LDSC}
\usage{
run_LDSC(exposure_data, exposure_name, outcome_data, outcome_name, ld, hm3,
  save_logfiles, verbose)
}
\arguments{
\item{exposure_name}{The name of the exposure trait, \code{default="exposure"} (character)}

\item{outcome_data}{}

\item{outcome_name}{The name of the outcome trait, \code{default="outcome"} (character)}

\item{ld}{The path to the folder in which the LD scores used in the analysis are located.
Expects LD scores formated as required by the original LD score regression software.  (character)}

\item{hm3}{The path to the folder in which the LD scores used in the analysis are located.
(character)}

\item{save_logfiles}{A logical indicating if log files from LDSC should be saved,
\code{default=FALSE}}

\item{verbose}{A logical indicating if information on progress should be reported,
\code{default=TRUE}}
}
\description{
Use GenomicSEM to perform cross-trait LDSC analysis and returns
the heritability of the exposure (SE) and the cross-trait intercept (SE)
}
