% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_MR.R
\name{run_MR}
\alias{run_MR}
\title{Run MR}
\usage{
run_MR(exposure_data, outcome_data, MR_threshold = 5e-08,
  MR_pruning_dist = 500, MR_pruning_LD = 0, MR_reverse = NULL,
  verbose = TRUE)
}
\arguments{
\item{outcome_data}{}

\item{MR_threshold}{The threshold used to select strong instruments for MR, should be lower
than 1e-5, \code{default=5e-8} (numeric)}

\item{MR_pruning_dist}{The distance used for pruning MR instruments (in Kb), should be between 10 and 1000,
\code{default=500} (numeric)}

\item{MR_pruning_LD}{The LD threshold used for pruning MR instruments, should be between 0 and 1
(if 0, distance-based pruning is used), \code{default=0} (numeric)}

\item{MR_reverse}{The ,\code{default=1e-3} (numeric)}

\item{verbose}{A logical indicating if information on progress should be reported,
\code{default=TRUE}}
}
\description{
Use TwoSampleMR to perform IVW-MR and returns causal effect estimate (SE)
but also M (number of instruments), mean sample size for exposure / outcome
and the set of IVs used
}
