% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_inputGWAS.R
\name{tidy_inputGWAS}
\alias{tidy_inputGWAS}
\title{Tidy input GWAS}
\usage{
tidy_inputGWAS(GWAS, verbose = FALSE)
}
\arguments{
\item{GWAS}{xx}

\item{verbose}{A logical indicating if information on progress should be reported,
\code{default=TRUE}}
}
\description{
From the GWAS arguments (exposure/outcome) of the main MRlap function,
create a nice/tidy data.frame that can be used by all other functions.
}
